\name{atmospheres}
\alias{atmospheres}
\title{atmospheres}

\description{
Atmospheres of Solar System.

Mercury might be easily taken out because it does not have atmosphere in strict sense.

All data in percentages.
}

\usage{atmospheres}

% \format{This data frame contains the following columns:
% \describe{
% \item{\code{...}}{...}
% }
% }

\source{
Data from the NASA Web site, once was available as 'planetatmoscomp.pdf' document.
}

% \examples{
% ...
% }

\keyword{datasets}
