\name{chaetocnema}
\alias{chaetocnema}

\title{Chaetocnema flea beetles}

\usage{chaetocnema}

\format{
These data frame contains the following columns:
\describe{
\item{\code{Species}}{Species epithet}
\item{\code{No}}{Number of sample (see below)}
\item{\code{x10}}{Width of the first joint of the first tarsus (the sum of measurements for both tarsi), in microns}
\item{\code{x12}}{The same for the second joint}
\item{\code{x14}}{The maximal width of the aedeagus in the fore-part, in microns}
\item{\code{x18}}{The front angle of the aedeagus, 1 unit = 7.5 degrees}
\item{\code{x40}}{The maximal width of the head between the external edges of the eyes, in 0.01 mm}
\item{\code{x48}}{The aedeagus width from the side, in microns}
}
}

\description{
Lubischew data (1962, pp. 465--468, tables 4--6): 74 Chaetocnema flea beetles specimens which belong to three cryptic species.

Sources of specimens:

Chaetocnema concinna Marsh:

1-6 Environs of Uljianovsk;
7 Khvalynsk, the Volga;
8-9 Perm;
10-14 Environs of Leningrad;
15-17 The Ukraine;
18 Ashkhabad, Turkmenistan;
19-21 France.

Ch. heikertintgeri Lubis.:

1-8 Environs of Uljianovsk;
9 Khvalynsk;
10-14 Perm;
15-17 Environs of Leningrad;
18-20 The Ukraine;
21 Ustj-Zilma;
22 Gagra, Abkhazia;
23-27 Ussuri district;
28-29 Yakutsk district;
30 Khabarovsk;
31 Germany.

Ch. heptapotamnica Lubis.:

1-18 Environs of Lake Issyk-Kul, Kirghizia;
19 Alma-ata, Kazachstan;
20-22 Environs of Frunze, Kirghizia.
}

\source{
Lubischew A.A. 1962. On the use of discriminant functions in taxonomy. Biometrics. 18:455--477.
}

\keyword{datasets}
