\name{dolbli}
\alias{dolbli}
\title{dolbli}

\description{
Plants of two Arctic lakes.

Observations on the coastal flora of Arctic lakes. Flora was sampled on the 20 m coastline. 1999--2004.
}

\usage{dolbli}

\format{
\describe{
\item{\code{columns}}{Lake names with plot numbers, data is abundance of plant species, in 1543 scale
(0 -- absent; 1 -- one individual plant; 2 -- no more than 12 individual plants (rametes);
3 -- number of individuals is more than 12 but no more than 5\% of total number of plants on a plot;
4 -- number of individuals is more than 5\% but no more than 25\% of total number of plants on a plot;
5 -- number of individuals is more than 25\% but no more than 50\% of total number of plants on a plot;
6 -- number of individuals is more than 50\% but no more than 75\% of total number of plants on a plot;
7 -- number of individuals is more than 75\% of total number of plants on a plot.)}
\item{\code{rows}}{Names of plant species, trees start with 0}
}
}

\source{
Shipunov, A. The creation of databases about flora of isles and lakes of North
Karelia. Abstract. P. 97--98. Study of the flora of East Europe: Achievements and prospects.
23--28 May 2005, St.-Petersburg.

Shipunov A., Motyleva M. The comparative study of the flora of lakes in Tchupa gulf environs.
III scientific session of Marine Biological Station of Saint-Petersburg State University.
Abstracts. P. 27--29. Saint-Petersburg, 2002. [In Russian].
}

% \examples{
% ...
% }

\keyword{datasets}
