\name{drosera}
\alias{drosera}

\title{drosera}

\description{
Observations on the round-leaf sundew, Drosera rotundifolia, White Sea coast, August 2000.
}

\usage{
drosera
}

\format{
This data frame contains the following columns:
\describe{
\item{\code{POP}}{Code of the population}
\item{\code{YOUNG.L}}{Number of young, not opened leaves}
\item{\code{MATURE.L}}{Number of mature, catching leaves}
\item{\code{OLD.L}}{Number of old, degrading leaves}
\item{\code{INSECTS}}{Total number of insects per plant}
\item{\code{INFL.L}}{Inflorescence length (0 if absent), mm}
\item{\code{STALK.L}}{Length of stalk (without flowers), mm}
\item{\code{N.FLOW}}{Number of flowers}
\item{\code{LEAF.L}}{Length of maximal leaf, mm}
\item{\code{LEAF.W}}{Width of maximal leaf, mm}
\item{\code{PET.L}}{Length of maximal leaf petiole, mm}
}
}

% \source{
% ...
% }

% \examples{
% ...
% }

\keyword{datasets}
