\name{eq}
\alias{eq}
\alias{eq_l}
\alias{eq_s}

\title{eq}

\description{
Horsetails (\emph{Equisetum}) are the old, pre-dinosaur plant lineage.
Only several dozen species survived, but despite a long evolution the
borders between these species are still unclear for researchers.

In 2005--2006, morphometric analysis was performed of more than 1,000 horsetail
plants belong to most widespread Eurasian species growing in Middle Russia.
For the analysis, we used 8 morphological characters and also tried to identify species.

'eq_l' contains population locations and species determinations.

'eq_s' and 'eq' are actual morphometric data, but 'eq' contains only 2 species out of 8.
}

\usage{
eq
eq_l
eq_s
}

\format{
This data frame contains the following columns:
\describe{
\item{\code{DL.R}}{plant height, mm}
\item{\code{DIA.ST}}{maximal diameter of stem, mm}
\item{\code{N.REB}}{number of ridges on a stem}
\item{\code{N.ZUB}}{number of teeth (reduced leaves)}
\item{\code{DL.OSN.Z}}{length of tooth base}
\item{\code{DL.TR.V}}{length of sheath}
\item{\code{DL.BAZ}}{length of basal segment of branch}
\item{\code{DL.PER}}{length of first (after the basal) segment of branch}
\item{\code{SPECIES}}{preliminary species determination}
\item{\code{N.POP}}{population number}
\item{\code{WHERE}}{population location (region)}
}}

% \source{
%  ...
% }

% \examples{
% ...
% }

\keyword{datasets}
