\name{Ex.boxplot}
\alias{Ex.boxplot}
\title{Boxplot explanation}
\description{Boxplot explanation}

\usage{Ex.boxplot(...)}

\arguments{
\item{...}{Arguments to 'boxplot()'}
}

\details{The scheme which explains typical boxplot.}

% \value{}

% \references{}

\author{Alexey Shipunov}

\seealso{\code{\link{boxplot}}}

\examples{
Ex.boxplot()
}

\keyword{hplot}

