\name{Ex.col}

\title{Examples of colors}

\alias{Ex.col}
\alias{Ex.cols}

\description{Examples of standard colors (also shows all colors)}

\usage{
Ex.col(all=FALSE)
Ex.cols(all=FALSE)
}

\arguments{
\item{all}{if TRUE, shows all 476 named colors}
}

\details{
Examples of standard colors (also shows all colors). For the palettes, run 'example(rainbow)'.

Note: large device is required to see all 476 named colors.
}

% \value{}

% \references{}

\author{Alexey Shipunov}

\seealso{\code{\link{palette}}}

\examples{
Ex.cols()
Ex.cols(all=TRUE)
}

\keyword{hplot}
