\name{Ex.font}
\title{Examples of fonts}
\alias{Ex.font}
\alias{Ex.fonts}

\description{Examples of standard fonts}

\usage{Ex.font()}

% \arguments{}

\details{Examples of standard fonts}

% \value{}

% \references{}

\author{Alexey Shipunov}

\seealso{\code{\link{par}}}

\examples{
Ex.fonts()
}

\keyword{hplot}

