\name{Ex.lty}
\alias{Ex.lty}
\alias{Ex.lines}

\title{Examples of line types}
\description{Line type examples}

\usage{
Ex.lty(custom="431313")
Ex.lines(custom="431313")
}

\arguments{

\item{custom}{character string to specify custom line type (see '?lines').}}

\details{Line type examples. To see other possible custom line types, try custom="F8" or similar.}

% \value{}

% \references{}

\author{Alexey Shipunov}

\seealso{\code{\link{lines}}}

\examples{
Ex.lines(custom="F8")
}

\keyword{hplot}

