\name{Ex.margins}
\alias{Ex.margins}
\title{Example of plot margins}

\description{Example of plot margins}

\usage{Ex.margins()}

% \arguments{}

\details{Example of plot margins. Modified from Paul Murrell (2006).}

% \value{}

\references{Murrell P. 2006. R Graphics.}

\author{Alexey Shipunov}

\seealso{\code{\link{par}}}

\examples{
Ex.margins()
}

\keyword{hplot}

