\name{Ex.plots}
\title{Examples of plot types}
\alias{Ex.plots}
\alias{Ex.types}

\description{Examples of plot types}

\usage{Ex.plots()
Ex.types()}

% \arguments{}

\details{Examples of nine standard plot types.}

% \value{}

% \references{}

\author{Alexey Shipunov}

\seealso{\code{\link{par}}}

\examples{
Ex.types()
}

\keyword{hplot}
