\name{Saynodynamite}
\alias{Saynodynamite}
\title{Say ``no'' to dynamite plots!}
\description{Say ``no'' to dynamite plots!}

\usage{Saynodynamite()}

% \arguments{}

\details{'Poster' plot to emphazise the harmfullness of so called dynamite plots. See, for example, 'http://emdbolker.wikidot.com/blog\%3Adynamite'.}

% \value{}

% \references{}

\author{Alexey Shipunov}

\seealso{\code{\link{boxplot}}}

\examples{
Saynodynamite()
}

\keyword{Graphical fun}

