\name{Miney}
\alias{Miney}
\title{Miney game}
\description{Miney minesweeper game}

\usage{Miney(n, ucol="black", gcol="white", bcol="red")}

\arguments{
\item{n}{indicates the size of the matrix to play, i.e. n=5 results in n x n = 5 x 5 matrix.}
\item{ucol}{color of unknown cells}
\item{gcol}{color of good cells}
\item{bcol}{color of bad cells}
}

\details{Miney game, modified from 'miney' package of Roland Rau. See also the 'fun' package.}

% \value{}

% \references{}

\author{Alexey Shipunov}

% \seealso{}

\examples{
\dontrun{
## interactive command:
Miney(3)
}
}

\keyword{Graphical fun}

