\name{Saynodynamite}
\alias{Saynodynamite}
\title{Say ``no'' to dynamite plots!}
\description{Say ``no'' to dynamite plots!}

\usage{Saynodynamite()}

% \arguments{}

\details{
'Poster' plot to emphasize the harmfulness of so called dynamite plots.

See, for example, thorough analysis at "http://emdbolker.wikidot.com/blog\%3Adynamite".
}

% \value{}

% \references{}

\author{Alexey Shipunov}

\seealso{\code{\link{boxplot}}}

\examples{
Saynodynamite()
}

\keyword{hplot}

